Add 'serverSentEvent+' and 'serverSentEvent+js'
This is the result of an half successful attempt to dynamically add
sse's to the app. 'serverSentEvent+' works as expected and can add
arbitrary HTML to an element (instead of replacing it).
'serverSentEvent+js' was an attempt to get the browser to add and
execute arbitrary scripts, to install new sse's on the fly. It does not
work. I have decided that this is not path forward.