Add 'make-rhythms', 'pack-rhythms', 'make-melody'
This is huge! It's a little DSL for quickly creating rhythmic patterns.
Tuplets are automatically generated. These rhythmic patterns can be
mapped to note generator functions, etc.

Create rhythmic patterns with 'make-rhythms', map the patterns to notes
with pack-rhythms, and convert it all to something lilypond can process
with 'make-melody'.

NOTE - I'm not sold on the name 'make-melody'. It will likely be used
more for creating arpeggios and other generated music. Maybe
'make-arpeggio'? 'make-music'? lol. I'll figure it out.